[Refactor] Move pages to its own directory
